requirements.txt
dj-rest-auth==2.2.4
Django==4.0.5
djangorestframework==3.13.1
drf-yasg==1.20.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-decouple==3.6
pillow==10.2.0